gestures: add knob to allow end-users to disable gestures
Currently, gesture recognizers are added to capture tap and pan events,
which assists in allowing users to swipe the view, and in turn run a
swipe animation.  It is also possible to set the allowable swiping
direction to .None which will prevent the top-view card from reaching a
threshold and moving to the following card.

However, it may be useful for some to not only disable the allowable
direction, but also disable the gesture recognizer to effectively
disable the animations.  The ViewManager is useful for stacking the
views, and if end-users may not want to allow users to swipe, disabling
the gesture recognizer is a convenient shortcut.

By default gestures are enabled since this feature would be largely
pointless without them.

Signed-off-by: Jamie Couture <jamie.couture@crowdspark.com>