updateViews(): drop gesture recognizer block short-circuit logic
Introduced since initial commit, when updating views there is an
optimization introduced to skip animating views when there are no
gestures found for the active view.  If users set `ignoreGestures` the
gesture recognizer collection will always be empty triggering the false
case of the guard statement and returning, prematurely skipping the
animating of the desired view.

After running through a few scenario tests, the only way to trigger the
path to fail the guard statement and return (short-circuit) is to not
define a gesture on any view.  In other words, this logic block does not
appear to reduce animation steps or avoid unnecessary calculations since
all our containing views will have a gesture to find, and the false case
will never be executed.  If callers choose to skip adding gestures to
avoid the swipe and tap gestures from being applied, but also want to
use history, we should not return early.

Fixes an issue where calling rewind() is broken and does not update the
active view when users set ignoreGestures to true.

Signed-off-by: Jamie Couture <jamie.couture@crowdspark.com>